% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekg.preg-package.R
\docType{package}
\name{ekg.preg-package}
\alias{ekg.preg}
\alias{ekg.preg-package}
\title{ekg.preg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
